Box's M-test

It performs the Box's M-test for homogeneity of covariance matrices obtained from multivariate normal data according to one classification factor. The test is based on the chi-square approximation.

boxM(data, grouping)

Arguments

data
a numeric data.frame or matrix containing n observations of p variables; it is expected that n > p.

grouping
a vector of length n containing the class of each observation; it is usualy a factor.

Value

A list with class "htest" containing the following components:

statistic

an approximated value of the chi-square distribution.

parameter

the degrees of freedom related of the test statistic in this case that it follows a Chi-square distribution.

p.value

the p-value of the test.

cov

a list containing the within covariance matrix for each level of grouping.

pooled

the pooled covariance matrix.

logDet

a vector containing the natural logarithm of each matrix in cov.

data.name

a character string giving the names of the data.

method

the character string "Box's M-test for Homogeneity of Covariance Matrices".

References

Morrison, D.F. (1976) Multivariate Statistical Methods.

Author

Anderson Rodrigo da Silva <anderson.agro@hotmail.com>

Examples

data(iris)
boxM(iris[, -5], iris[, 5])

#> 
#> 	Box's M-test for Homogeneity of Covariance Matrices
#> 
#> data:  iris[, -5]
#> Chi-Sq (approx.) = 140.94, df = 20, p-value < 2.2e-16
#> 


# End (not run)