Pairwise Correlation t-Test

It performs multiple correlation t-tests from a correlation matrix based on the statistic: $$t = r * \sqrt(df / (1 - r^2))$$ where, in general, \(df = n - 2\).

multcor.test(x, n = NULL, Df = NULL,
  alternative = c("two.sided", "less", "greater"), adjust = "none")

Arguments

x
a correlation matrix.

n
the number of observations; if NULL (default), the argument Df must be passed.

Df
the number of degrees of freedom of the t statistic; if NULL (default), the argument n must be passed and, in this case, multcor.test considers \(Df = n - 2\).

alternative
the alternative hypothesis. It must be one of "two.sided", "greater" or "less". You can specify just the initial letter. "greater" corresponds to positive association,"less" to negative association. The default is "two.sided".

adjust
The adjustment method for multiple tests. It must be one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none" (default). For more information, see p.adjust.

Value

A list with class "multcor.test" containing the following components:

t.values

the t-value calculated for each correlation.

p.values

the p.value for each t-test, adjusted for multiple tests.

p.check

a matrix containing the p.values for each t-test (lower triangular) and a symbol indicating the significance level at which one can to reject the null hypothesis (upper triangular).

adjustemnt

a character indicating the p-value adjustment method.

df

the degrees of freedom of the tests.

alternative

a character indicating the type of alternative hypothesis.

data.name

a character string giving the name of the data.

Author

Anderson Rodrigo da Silva <anderson.agro@hotmail.com>

See also

Examples

data(peppercorr)
multcor.test(peppercorr, n = 20)

#> 
#> 	Pairwise correlation t-test
#> 
#> data:  peppercorr
#> degrees of freedom: 18 
#> alternative hypothesis: the true correlation is not equal to 0 
#> p-values (with none adjustment for multiple tests): 
#>     PL     FL     LFD   PT    AFW   DMC 
#> PL  <NA>          **                *   
#> FL  0.4676 <NA>                     *   
#> LFD 0.0035 0.9957 <NA>  ***   ***       
#> PT  0.1313 0.732  0.001 <NA>            
#> AFW 4e-04  0.5363 0     1e-04 <NA>      
#> DMC 0.0432 0.032  2e-04 5e-04 3e-04 <NA>
#> --- 
#> Signif. codes: '***'0.001 '**'0.01 '*'0.05 '.'0.1 ' '1 
#> 


# End (not run)